# architecture/campotech-openapi-spec.yaml
...
openapi: 3.1.0
info:
  title: CampoTech API
  description: |
    # CampoTech Field Service Management API
    
    Complete REST API for managing field service operations in Argentina.
    
    ## Features
    - Customer management with CUIT validation
    - Job scheduling and tracking
    - AFIP electronic invoicing (Factura Electrónica)
    - Mercado Pago payment processing
    - WhatsApp communication
    - Voice AI message processing
    
    ## Authentication
    All endpoints require Bearer token authentication except `/auth/*` endpoints.
    
    ## Rate Limits
    - Standard: 100 requests/minute per organization
    - Auth endpoints: 10 requests/minute per IP
    - WhatsApp: 50 messages/minute per organization
    - Voice processing: 20 requests/minute per organization
    
    ## Pagination
    List endpoints support cursor-based pagination with `limit` and `cursor` parameters.
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format.
    
  version: 1.0.0
  contact:
    name: CampoTech API Support
    email: api@campotech.com
    url: https://docs.campotech.com
  license:
    name: Proprietary
    url: https://campotech.com/terms
servers:
  - url: https://api.campotech.com/v1
    description: Production server
  - url: https://api.staging.campotech.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/auth/otp/verify`.
        Token contains: `sub` (user_id), `org_id`, `role`, `exp`.
        Access tokens expire in 15 minutes.
        Use refresh token to obtain new access token.

  # ==========================================================================
  # HEADERS
  # ==========================================================================
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: "CUSTOMER_001"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid CUIT format"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        field:
          type: string
          description: Field that caused the error (for validation errors)
          example: "cuit"
...
    UserRole:
      type: string
      enum:
        - owner
        - admin
        - technician
      description: |
        User role in organization:
        - `owner`: Full access including billing
        - `admin`: Operational access, no billing
        - `technician`: Own jobs only
...
    User:
      type: object
      required:
        - id
        - orgId
        - role
        - fullName
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/UserRole'
        fullName:
          type: string
          example: "Juan González"
        phone:
          type: string
          example: "+5491112345678"
        email:
          type: string
          format: email
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        lastSeenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    UserMe:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            organization:
              $ref: '#/components/schemas/OrganizationSummary'
            permissions:
              type: array
              items:
                type: string
              example: ["jobs.read", "jobs.write", "customers.read"]
    
    # --------------------------------------------------------------------------
    # Organization Schemas
    # --------------------------------------------------------------------------
...
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        cuit:
          type: string
        afipConfigured:
          type: boolean
        mpConnected:
          type: boolean
        whatsappConnected:
          type: boolean
...
paths:
...
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user
      description: Returns the authenticated user's profile and organization
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/UserMe'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==========================================================================
  # ORGANIZATION ENDPOINTS
  # ==========================================================================
...